feat : ProductCode 관련 엔티티 삭제 및 패키지 이름 변경
- 현재 데이터와의 호환성을 위해 코드 관련 엔티티 삭제 -> Varchar(15) 컬럼으로 변경
- 비고, 위치 관련 값 변경 도메인 로직 추가
- 패키지 이름 item -> product로 변경
