+++ src/main/java/dau/azit/simon/product/domain/ProductId.java
@@ -1,0 +1,12 @@
+package dau.azit.simon.product.domain;
+
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.GeneratedValue;
+
+import java.io.Serializable;
+
+@Embeddable
+public class ProductId implements Serializable {
+    @GeneratedValue
+    private Long id;
+}
